Clean up codebase structure: move submodule, remove Float type, clean debug code

+++ crates/libspot/src/config.rs
@@ -1,19 +1,16 @@
 //! Configuration types for SPOT detector
 
-use crate::Float;
-
-
 /// Configuration parameters for SPOT detector
 #[derive(Debug, Clone, PartialEq)]
 pub struct SpotConfig {
     /// Anomaly probability threshold (must be between 0 and 1-level)
-    pub q: Float,
+    pub q: f64,
     /// Whether to observe lower tail (false = upper tail, true = lower tail)
     pub low_tail: bool,
     /// Whether to discard anomalies from model updates
     pub discard_anomalies: bool,
     /// Excess level - high quantile that delimits the tail (must be between 0 and 1)
-    pub level: Float,
+    pub level: f64,
     /// Maximum number of excess data points to keep
     pub max_excess: usize,
 }
